// frameworks/bridge/arkts_frontend/koala_mirror/arkoala-arkts/arkui/src/generated/arkui-custom.ts
export class SceneInternal {
    public static fromPtr(ptr: KPointer): Scene {
        const obj : Scene = new Scene()
        obj.peer = new Finalizable(ptr, Scene.getFinalizer())
        return obj
    }
}
export class Scene implements MaterializedBase {
    peer?: Finalizable | undefined = undefined
    public getPeer(): Finalizable | undefined {
        return this.peer
    }
    static ctor_scene(): KPointer {
        const retval  = ArkUIGeneratedNativeModule._Scene_ctor()
        return retval
    }
    constructor() {
        // Constructor does not have parameters.
        // It means that the static method call invokes ctor method as well
        // when all arguments are undefined.
        const ctorPtr : KPointer = Scene.ctor_scene()
        this.peer = new Finalizable(ctorPtr, Scene.getFinalizer())
    }
    static getFinalizer(): KPointer {
        return ArkUIGeneratedNativeModule._Scene_getFinalizer()
    }
    public static load(uri?: ResourceStr): Scene {
        const uri_casted = uri as (ResourceStr | undefined)
        return Scene.load_serialize(uri_casted)
    }
    public destroy(): void {
        this.destroy_serialize()
        return
    }
    private static load_serialize(uri?: ResourceStr): Scene {
        const thisSerializer : Serializer = Serializer.hold()
        let uri_type : int32 = RuntimeType.UNDEFINED
        uri_type = runtimeType(uri)
        thisSerializer.writeInt8(uri_type as int32)
        if ((RuntimeType.UNDEFINED) != (uri_type)) {
            const uri_value  = uri!
            let uri_value_type : int32 = RuntimeType.UNDEFINED
            uri_value_type = runtimeType(uri_value)
            if (RuntimeType.STRING == uri_value_type) {
                thisSerializer.writeInt8(0 as int32)
                const uri_value_0  = uri_value as string
                thisSerializer.writeString(uri_value_0)
            }
            else if (RuntimeType.OBJECT == uri_value_type) {
                thisSerializer.writeInt8(1 as int32)
                const uri_value_1  = uri_value as Resource
                thisSerializer.writeResource(uri_value_1)
            }
        }
        const retval  = ArkUIGeneratedNativeModule._Scene_load(thisSerializer.asBuffer(), thisSerializer.length())
        thisSerializer.release()
        const obj : Scene = SceneInternal.fromPtr(retval)
        return obj
    }
    private destroy_serialize(): void {
        ArkUIGeneratedNativeModule._Scene_destroy(this.peer!.ptr)
    }
}